fix/refatoração mensagem
