+++ src/components/Chat/ChatMessages/ChatMessages.jsx
@@ -1,106 +1,23 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable react/prop-types */
-import { useState, useRef, useEffect, useContext } from "react";
+import { useContext } from "react";
 import { DownloadSimple, Checks, CaretDown } from "@phosphor-icons/react";
 import ModalMessage from "../Modals/ModalMessage";
-import { ChatContext } from "../../../context/ChatContext";
-import { ControlContext } from "../../../context/ControlContext";
+import { MessageContext } from "../../../context/MessageContext";
 
 
 const ChatMessages = () => {
 
   const {
-    highlightText, 
-    searchMessage, 
-    highlightedMessageIndex 
-  } = useContext(ChatContext);
+    highlightText,
+    searchMessage,
+    highlightedMessageIndex,
+    filteredMessages,
+    messagesEndRef,
+    isModalOpen,
+    openModal,
 
-  const {
-    selectedContact,
-    contactsMessages,
-  } = useContext(ControlContext);
-
-
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [currentMessage, setCurrentMessage] = useState(null);
-  const [modalPosition, setModalPosition] = useState({ top: 0, left: 0 });
-  const messagesEndRef = useRef(null);
-
-
-  const filteredMessages = contactsMessages[selectedContact]
-    ? contactsMessages[selectedContact].map((msg, index) => ({
-        ...msg,
-        index,
-      }))
-    : [];
-
-
-  useEffect(() => {
-    scrollToBottom();
-  }, [filteredMessages]);
-
-  const scrollToBottom = () => {
-    messagesEndRef.current?.scrollIntoView({ behavior: "smooth" });
-  };
-
-  const scrollToHighlightedMessage = (index) => {
-    const highlightedMessage = document.querySelector(`.message[data-index="${index}"]`);
-    if (highlightedMessage) {
-      highlightedMessage.scrollIntoView({ behavior: "smooth", block: "center" });
-    }
-  };
-
-  useEffect(() => {
-    if (highlightedMessageIndex !== -1) {
-      scrollToHighlightedMessage(highlightedMessageIndex);
-    }
-  }, [highlightedMessageIndex]);
-
-  const openModal = (message, event) => {
-    const rect = event.target.getBoundingClientRect();
-    const windowWidth = window.innerWidth;
-    const modalWidth = 150; // Largura aproximada do modal
-
-    let left = rect.left + window.scrollX;
-
-    if (windowWidth - rect.right < modalWidth) {
-      left = rect.left + window.scrollX - modalWidth;
-    }
-
-    setModalPosition({ top: rect.top + window.scrollY, left });
-    setCurrentMessage(message);
-    setIsModalOpen(true);
-  };
-
-  const closeModal = () => {
-    setIsModalOpen(false);
-    setCurrentMessage(null);
-  };
-
-  const handleReply = () => {
-    console.log("Reply to:", currentMessage);
-    closeModal();
-  };
-
-  const handleForward = () => {
-    console.log("Forward:", currentMessage);
-    closeModal();
-  };
-
-  const handleCopy = () => {
-    navigator.clipboard.writeText(currentMessage.content);
-    closeModal();
-  };
-
-  const handleEdit = () => {
-    navigator.clipboard.writeText(currentMessage.content);
-    closeModal();
-  };
-
-  const handleDelete = () => {
-    console.log("Delete:", currentMessage);
-    closeModal();
-  };
+  } = useContext(MessageContext);
 
   return (
     <div className="bg-gray-200 flex-grow p-4 overflow-y-auto h-32">
@@ -141,15 +58,7 @@ const ChatMessages = () => {
         <div ref={messagesEndRef} />
       </div>
       {isModalOpen && (
-        <ModalMessage
-          onClose={closeModal}
-          onReply={handleReply}
-          onForward={handleForward}
-          onEdit={handleEdit}
-          onCopy={handleCopy}
-          onDelete={handleDelete}
-          position={modalPosition}
-        />
+        <ModalMessage/>
       )}
     </div>
   );
