+++ src/components/Chat/Modals/ModalMessage.jsx
@@ -1,18 +1,34 @@
 /* eslint-disable react/prop-types */
-const ModalMessage = ({ onClose, onReply, onForward, onEdit, onCopy, onDelete, position }) => {
+import { MessageContext } from '../../../context/MessageContext';
+import { useContext } from 'react';
+
+const ModalMessage = () => {
+
+  const {
+    closeModal,
+    handleReply,
+    handleForward,
+    handleCopy,
+    handleEdit,
+    handleDelete,
+    modalPosition,
+
+  } = useContext(MessageContext);
+
+
   return (
-    <div className="fixed inset-0 z-20 flex items-center justify-center" onClick={onClose}>
+    <div className="fixed inset-0 z-20 flex items-center justify-center" onClick={closeModal}>
       <div
         className="bg-white p-4 rounded shadow-md"
-        style={{ position: "absolute", top: position.top, left: position.left }}
+        style={{ position: "absolute", top: modalPosition.top, left: modalPosition.left }}
         onClick={(e) => e.stopPropagation()} // Prevent closing when clicking inside the modal
       >
         <ul>
-          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={onReply}>Reply</li>
-          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={onForward}>Forward</li>
-          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={onCopy}>Copy</li>
-          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={onEdit}>Edit</li>
-          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={onDelete}>Delete</li>
+          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={handleReply}>Reply</li>
+          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={handleForward}>Forward</li>
+          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={handleCopy}>Copy</li>
+          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={handleEdit}>Edit</li>
+          <li className="py-2 px-4 hover:bg-gray-200 cursor-pointer" onClick={handleDelete}>Delete</li>
         </ul>
       </div>
     </div>
